ai-hr/services/tts/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
piper-phonemize==1.1.0
numpy==2.0.1
python-dotenv==1.0.1